+++ packages/react-router/src/index.ts
@@ -6,5 +6,9 @@ export {
   withRouter,
   RouterChildContext,
   MemoryRouter,
+  useRouteMatch,
+  useParams,
+  useLocation,
+  useHistory,
 } from 'react-router-dom';
 export {Router, Redirect, Link} from './components';