+++ packages/koa-shopify-webhooks/src/register.ts
@@ -8,8 +8,14 @@ export enum ApiVersion {
   October19 = '2019-10',
   January20 = '2020-01',
   April20 = '2020-04',
+  July20 = '2020-07',
   Unstable = 'unstable',
   Unversioned = 'unversioned',
+}
+
+export enum DeliveryMethod {
+  Http = 'http',
+  EventBridge = 'eventbridge',
 }
 
 export interface Options {
@@ -18,6 +24,7 @@ export interface Options {
   shop: string;
   accessToken: string;
   apiVersion: ApiVersion;
+  deliveryMethod: DeliveryMethod;
 }
 
 export async function registerWebhook({
@@ -26,12 +33,13 @@ export async function registerWebhook({
   accessToken,
   shop,
   apiVersion,
+  deliveryMethod = DeliveryMethod.Http,
 }: Options) {
   const response = await fetch(
     `https://${shop}/admin/api/${apiVersion}/graphql.json`,
     {
       method: Method.Post,
-      body: buildQuery(topic, address),
+      body: buildQuery(topic, address, deliveryMethod),
       headers: {
         [WebhookHeader.AccessToken]: accessToken,
         [Header.ContentType]: 'application/graphql',
@@ -41,21 +49,46 @@ export async function registerWebhook({
 
   const result = await response.json();
 
-  if (
-    result.data &&
-    result.data.webhookSubscriptionCreate &&
-    result.data.webhookSubscriptionCreate.webhookSubscription
-  ) {
-    return {success: true, result};
-  } else {
-    return {success: false, result};
+  return {success: isSuccess(result, deliveryMethod), result};
+}
+
+function isSuccess(result, deliveryMethod: DeliveryMethod): boolean {
+  switch (deliveryMethod) {
+    case DeliveryMethod.Http:
+      return Boolean(
+        result.data &&
+          result.data.webhookSubscriptionCreate &&
+          result.data.webhookSubscriptionCreate.webhookSubscription,
+      );
+    case DeliveryMethod.EventBridge:
+      return Boolean(
+        result.data &&
+          result.data.eventBridgeWebhookSubscriptionCreate &&
+          result.data.eventBridgeWebhookSubscriptionCreate.webhookSubscription,
+      );
   }
 }
 
-function buildQuery(topic: string, callbackUrl: string) {
+function buildQuery(
+  topic: string,
+  address: string,
+  deliveryMethod: DeliveryMethod,
+) {
+  let mutationName;
+  let webhookSubscriptionArgs;
+  switch (deliveryMethod) {
+    case DeliveryMethod.Http:
+      mutationName = 'webhookSubscriptionCreate';
+      webhookSubscriptionArgs = `{callbackUrl: "${address}"}`;
+      break;
+    case DeliveryMethod.EventBridge:
+      mutationName = 'eventBridgeWebhookSubscriptionCreate';
+      webhookSubscriptionArgs = `{arn: "${address}"}`;
+      break;
+  }
   return `
     mutation webhookSubscriptionCreate {
-      webhookSubscriptionCreate(topic: ${topic}, webhookSubscription: {callbackUrl: "${callbackUrl}"}) {
+      ${mutationName}(topic: ${topic}, webhookSubscription: ${webhookSubscriptionArgs}) {
         userErrors {
           field
           message