+++ packages/react-import-remote/src/index.ts
@@ -1,3 +1,5 @@
 import ImportRemote from './ImportRemote';
 
+export {useImportRemote, Status} from './hooks';
+
 export default ImportRemote;