+++ packages/react-router/src/components/Link/Link.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import {Link as ReactRouterLink} from 'react-router-dom';
+import {Link as ReactRouterLink, LinkProps} from 'react-router-dom';
 
-export interface Props {
+type ReactRouterLinkProps = Omit<LinkProps, 'to'>;
+
+export interface Props extends ReactRouterLinkProps {
   external?: boolean;
   url: string;
-  children?: React.ReactNode;
 }
 
 export default function Link({url, external, children, ...rest}: Props) {