+++ packages/react-router/src/index.ts
@@ -5,5 +5,10 @@ export {
   RouteComponentProps,
   withRouter,
   RouterChildContext,
+  MemoryRouter,
+  useRouteMatch,
+  useParams,
+  useLocation,
+  useHistory,
 } from 'react-router-dom';
 export {Router, Redirect, Link} from './components';