+++ packages/react-server/src/server/server.ts
@@ -5,11 +5,10 @@ import Koa, {Context} from 'koa';
 import compose from 'koa-compose';
 import mount from 'koa-mount';
 
-import {createRender, RenderFunction} from '../render';
+import {createRender, RenderFunction, RenderOptions} from '../render';
 import {requestLogger} from '../logger';
 import {metricsMiddleware as metrics} from '../metrics';
 import {ping} from '../ping';
-import {ValueFromContext} from '../types';
 
 const logger = console;
 
@@ -17,10 +16,13 @@ interface Options {
   ip?: string;
   port?: number;
   assetPrefix?: string;
-  assetName?: string | ValueFromContext<string>;
+  proxy?: boolean;
+  assetName?: RenderOptions['assetName'];
+  htmlProps?: RenderOptions['htmlProps'];
   serverMiddleware?: compose.Middleware<Context>[];
   render: RenderFunction;
-  renderError?: RenderFunction;
+  renderError?: RenderOptions['renderError'];
+  app?: Koa;
 }
 
 /**
@@ -31,19 +33,29 @@ interface Options {
 export function createServer(options: Options): Server {
   const {
     /* eslint-disable no-process-env */
-    ip = process.env.REACT_SERVER_IP || 'localhost',
-    port = (process.env.REACT_SERVER_PORT &&
-      parseInt(process.env.REACT_SERVER_PORT, 10)) ||
-      8081,
+    ip = process.env.REACT_SERVER_IP &&
+    process.env.REACT_SERVER_IP !== 'undefined'
+      ? process.env.REACT_SERVER_IP
+      : '0.0.0.0',
+    port = process.env.REACT_SERVER_PORT &&
+    process.env.REACT_SERVER_PORT !== 'undefined'
+      ? parseInt(process.env.REACT_SERVER_PORT, 10)
+      : 8081,
     // a default is set in sewingKitMiddleware
-    assetPrefix = process.env.CDN_URL,
+    assetPrefix = process.env.CDN_URL && process.env.CDN_URL !== 'undefined'
+      ? process.env.CDN_URL
+      : undefined,
     /* eslint-enable no-process-env */
     render,
+    renderError,
     serverMiddleware,
     assetName,
-    renderError,
+    htmlProps,
+    proxy = false,
+    app = new Koa(),
   } = options;
-  const app = new Koa();
+
+  app.proxy = proxy;
 
   app.use(mount('/services/ping', ping));
 
@@ -54,9 +66,11 @@ export function createServer(options: Options): Server {
     app.use(compose(serverMiddleware));
   }
 
-  app.use(createRender(render, {assetPrefix, assetName, renderError}));
+  app.use(
+    createRender(render, {assetPrefix, assetName, renderError, htmlProps}),
+  );
 
-  return app.listen(port || 3000, ip, () => {
+  return app.listen(port, ip, () => {
     logger.log(`started react-server on ${ip}:${port}`);
   });
 }