+++ packages/react-server/src/webpack-plugin/webpack-plugin.ts
@@ -19,12 +19,14 @@ export class ReactServerPlugin {
     port,
     assetPrefix,
     basePath = '.',
+    proxy = false,
   }: Partial<Options> = {}) {
     this.options = {
       basePath,
       host,
       port,
       assetPrefix,
+      proxy,
     };
   }
 
@@ -58,7 +60,7 @@ export class ReactServerPlugin {
 }
 
 function serverSource(options: Options, compiler: Compiler) {
-  const {port, host, assetPrefix} = options;
+  const {port, host, assetPrefix, proxy} = options;
 
   return `
     ${HEADER}
@@ -92,12 +94,13 @@ function serverSource(options: Options, compiler: Compiler) {
       port: ${port},
       ip: ${JSON.stringify(host)},
       assetPrefix: ${JSON.stringify(assetPrefix)},
+      proxy: ${proxy},
       render,
       ${
         errorSSRComponentExists(options, compiler)
           ? `renderError: (ctx) => {
               return React.createElement(Error, {
-                url: ctx.request.URL,
+                url: ctx.request.url,
                 data: ctx.state.quiltData,
                 error: ctx.state.quiltError,
               });